[refactoring] Removing meaningless setup task for JTS
There is no worthy to have the task defined or used as JTS for being set
needs restart of server which is not done either here or even under
basic set of tests.
Setting of JTS and orb is done by XSLT template when jboss as is copied
to target directory.