[JBTM-3106] fixing test failures on Narayana CDI changes
it happens to have issues to inject the Narayana CDI interceptor when
the tests forces using the internal org.jboss.jts module.
This is not needed as we use the WFTC project now and we don't need
to use the Narayana classes directly in the tests.
As Narayana CDI extension is not used then the CDI was not loaded
and happens the CDI was not capable to find it

[ERROR] org.jboss.as.test.integration.ejb.transaction.cmt.timeout.TransactionTimeoutTestCase  Time elapsed: 0.175 s  <<< ERROR!
org.jboss.arquillian.container.spi.client.container.DeploymentException:
Cannot deploy test-ejb-cmt-timeout.jar: {"WFLYCTL0062: Composite operation failed and was rolled back. Steps that failed:" => {"Operation step-1" => {"WFLYCTL0080: Failed services" => {"jboss.deployment.unit.\"test-ejb-cmt-timeout.jar\".WeldStartService" => "Failed to start service
    Caused by: org.jboss.weld.exceptions.DeploymentException: WELD-001408: Unsatisfied dependencies for type TransactionExtension with qualifiers @Default
  at injection point [BackedAnnotatedField] @Inject private com.arjuna.ats.jta.cdi.transactional.TransactionalInterceptorBase.extension
  at com.arjuna.ats.jta.cdi.transactional.TransactionalInterceptorBase.extension(TransactionalInterceptorBase.java:0)
"}}}}